using rust-script
