--- compare_benchmarks/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "compare_benchmarks"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-prettytable-rs = "^0.10"
-serde_json = "1.0"
-serde = { version = "1", features = ["derive"] }
