--- compare_benchmarks/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod parse_json;
-
-pub use parse_json::*;
